studywithtufts/requirements.txt
Django==2.2.6
django-appconf==1.0.3
django-imagekit==4.0.2
pilkit==2.0
Pillow==6.2.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0